+++ ncarview/src/lib/libctrans/translate.h
@@ -1,6 +1,8 @@
 #ifndef	_translate_
 #define	_translate_
 
+#include <stdlib.h>
+
 /*
  */
 #define	DEVWIN_LLX	0
@@ -26,14 +28,14 @@
 /*
  * Macros to convert VDC x and y's into device x and y's
  */
-#define	XScale(x)	(long) (labs(X_Scale * (x)))
-#define	YScale(y)	(long) (labs(Y_Scale * (y)))
+#define	XScale(x)	(long) (labs((long) (X_Scale * (x))))
+#define	YScale(y)	(long) (labs((long) (Y_Scale * (y))))
 
 /*
  * Macros to scale a device dimension to VDC dimension
  */
-#define	XScale_(x_)	(long) (labs(((double) (x_)) / X_Scale ))
-#define	YScale_(y_)	(long) (labs(((double) (y_)) / Y_Scale ))
+#define	XScale_(x_)	(long) (labs((long) (((double) (x_)) / X_Scale )))
+#define	YScale_(y_)	(long) (labs((long) (((double) (y_)) / Y_Scale )))
 
 /*
  *	structure to define a coordinate system
